Added `de/tuples` module
